requirements.txt
request
pandas
matplotlib
numpy
json